Add a retry to the startup detection integration test
On GitHub Actions particularly, this can be a bit fragile, with
browsers occasionally not being available. I've seen this once
or twice in real-world usage, I think because browser-launcher
detection can be a little slow. A retry to let it catch up
should solve that for testing use cases, and a couple of seconds
on first startup isn't a huge problem the rest of the time.