Update global-agent to hugely shrink JS override dep size
This removes the core-js dependency, changing nothing but shrinking the
overall size a long way (it's 85% of global-agent's install size), which
helps to speed up the Docker injection step significantly, and of course
shrinks the server itself nicely too. Every MB counts!

The remaining weight here is mostly from lodash, via global-tunnel-ng.
It is actually used there, could be shrunk with small changes, but the
module is unmaintained, so it's unlikely that's easy to do. Could be
worth forking in future, but for now it's acceptable, and we can
probably drop global-tunnel-ng entirely eventually, since it's only
required for Node < 10, which is entirely officually unsupported
versions already.