Ensure we always detect the correct host IP (especially on Linux)
For Mac & Windows, it appears that we can always use
docker.host.internal everywhere and we're all good (though this may
change with podman in future...).

On Linux though, this never works. We have to map that name to an IP
ourselves, and the IP isn't clearly predictable. We now:

* Use host-gateway where possible, an official alias that does the
correct mapping automatically.
* Use the existing container's network config to get the gateway IP
(_usually_ the host, but tbc) and use that.
* Use the default docker bridge network host IP, if all else fails.